Refactoring of LocalMUCRoom's 'joinRoom' method
Splitting up the large, complex method 'joinRoom' in LocalMUCRoom to have a better separation of responsibilities and more manageble code.

Added human readible text to most of the errors generated in this class.

Introduced the @Nullable and @Nonnull annotations to aid IDEs with problem detection.

Added lots of debug logging.

This commit should not have introduced significant functional changes.